Assignments_01/requirements.txt
streamlit
requests
pandas
plotly
pillow
matplotlib
python-dotenv
statsmodels
